@@ -1,21 +1,26 @@
 name: clima
-description: A new Flutter application.
+description: "Clima App for Flutter UI Design"
+publish_to: 'none'
 
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ^3.3.1
 
 dependencies:
   flutter:
     sdk: flutter
 
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.6
+  geolocator: ^11.0.0
+  http: ^1.2.1
+  flutter_spinkit: ^5.2.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
+  flutter_lints: ^3.0.0
+  change_app_package_name: ^1.1.0
 flutter:
   uses-material-design: true
 
